chore(run-protocol): consolidate attestation files
